Revert "rename filereader.xx->logreader.xx"
This reverts commit 75fc54f1e219c066e1d78c15284171131181c0d8.
